move QueryResult to protocol crate
