+++ src/node/src/node.rs
@@ -13,10 +13,9 @@
 // limitations under the License.
 
 use crate::query_listener::SmolQueryListener;
-use protocol::{listener::Secure, messages::Message, ColumnMetadata, Command, QueryListener};
+use protocol::{listener::Secure, Command, QueryListener};
 use smol::Task;
-use sql_engine::{Handler, QueryEvent, QueryResult};
-use sql_types::SqlType;
+use sql_engine::Handler;
 use std::sync::{
     atomic::{AtomicU8, Ordering},
     Arc, Mutex,
@@ -93,7 +92,7 @@ impl Node {
                             }
                             Ok(Ok(Command::Query(sql_query))) => {
                                 let response = sql_handler.execute(sql_query.as_str()).expect("no system error");
-                                match connection.send(QueryResultMapper::map(response)).await {
+                                match connection.send(response).await {
                                     Ok(()) => {}
                                     Err(error) => eprintln!("{:?}", error), // break Err(SystemError::io(error)),
                                 }
@@ -104,275 +103,5 @@ impl Node {
                 .detach();
             }
         });
-    }
-}
-
-struct TypeConverter;
-
-impl TypeConverter {
-    fn pg_oid(sql_type: &SqlType) -> i32 {
-        match sql_type {
-            SqlType::Bool => 16,
-            SqlType::Char(_) => 18,
-            SqlType::BigInt => 20,           // PG int8
-            SqlType::SmallInt => 21,         // PG int2
-            SqlType::Integer => 23,          // PG int4
-            SqlType::Real => 700,            // PG float4
-            SqlType::DoublePrecision => 701, // PG float8
-            SqlType::VarChar(_) => 1043,
-            SqlType::Date => 1082,
-            SqlType::Time => 1083,
-            SqlType::Timestamp => 1114,
-            SqlType::TimestampWithTimeZone => 1184, // PG Timestamptz
-            SqlType::Interval => 1186,
-            SqlType::TimeWithTimeZone => 1266, // PG Timetz
-            SqlType::Decimal => 1700,          // PG Numeric & Decimal
-        }
-    }
-
-    fn pg_len(sql_type: &SqlType) -> i16 {
-        match sql_type {
-            SqlType::Bool => 1,
-            SqlType::Char(_) => 1,
-            SqlType::BigInt => 8,
-            SqlType::SmallInt => 2,
-            SqlType::Integer => 4,
-            SqlType::Real => 4,
-            SqlType::DoublePrecision => 8,
-            SqlType::VarChar(_) => -1,
-            SqlType::Date => 4,
-            SqlType::Time => 8,
-            SqlType::Timestamp => 8,
-            SqlType::TimestampWithTimeZone => 8,
-            SqlType::Interval => 16,
-            SqlType::TimeWithTimeZone => 12,
-            SqlType::Decimal => -1,
-        }
-    }
-}
-
-struct QueryResultMapper;
-
-impl QueryResultMapper {
-    fn map(resp: QueryResult) -> Vec<Message> {
-        match resp {
-            Ok(QueryEvent::SchemaCreated) => vec![Message::CommandComplete("CREATE SCHEMA".to_owned())],
-            Ok(QueryEvent::SchemaDropped) => vec![Message::CommandComplete("DROP SCHEMA".to_owned())],
-            Ok(QueryEvent::TableCreated) => vec![Message::CommandComplete("CREATE TABLE".to_owned())],
-            Ok(QueryEvent::TableDropped) => vec![Message::CommandComplete("DROP TABLE".to_owned())],
-            Ok(QueryEvent::VariableSet) => vec![Message::CommandComplete("SET".to_owned())],
-            Ok(QueryEvent::TransactionStarted) => vec![Message::CommandComplete("BEGIN".to_owned())],
-            Ok(QueryEvent::RecordsInserted(records)) => vec![Message::CommandComplete(format!("INSERT 0 {}", records))],
-            Ok(QueryEvent::RecordsSelected(projection)) => {
-                let definition = projection.0;
-                let description: Vec<ColumnMetadata> = definition
-                    .into_iter()
-                    .map(|(name, sql_type)| {
-                        ColumnMetadata::new(name, TypeConverter::pg_oid(&sql_type), TypeConverter::pg_len(&sql_type))
-                    })
-                    .collect();
-                let records = projection.1;
-                let len = records.len();
-                let mut messages = vec![Message::RowDescription(description)];
-                for record in records {
-                    messages.push(Message::DataRow(record));
-                }
-                messages.push(Message::CommandComplete(format!("SELECT {}", len)));
-                messages
-            }
-            Ok(QueryEvent::RecordsUpdated(records)) => vec![Message::CommandComplete(format!("UPDATE {}", records))],
-            Ok(QueryEvent::RecordsDeleted(records)) => vec![Message::CommandComplete(format!("DELETE {}", records))],
-            Err(query_error) => vec![Message::ErrorResponse(
-                query_error.severity(),
-                query_error.code(),
-                Some(format!("{}", query_error)),
-            )],
-        }
-    }
-}
-
-#[cfg(test)]
-mod mapper {
-    use super::*;
-    use sql_engine::QueryError;
-    use sql_types::SqlType;
-
-    #[test]
-    fn create_schema() {
-        assert_eq!(
-            QueryResultMapper::map(Ok(QueryEvent::SchemaCreated)),
-            vec![Message::CommandComplete("CREATE SCHEMA".to_owned())]
-        )
-    }
-
-    #[test]
-    fn drop_schema() {
-        assert_eq!(
-            QueryResultMapper::map(Ok(QueryEvent::SchemaDropped)),
-            vec![Message::CommandComplete("DROP SCHEMA".to_owned())]
-        )
-    }
-
-    #[test]
-    fn create_table() {
-        assert_eq!(
-            QueryResultMapper::map(Ok(QueryEvent::TableCreated)),
-            vec![Message::CommandComplete("CREATE TABLE".to_owned())]
-        );
-    }
-
-    #[test]
-    fn drop_table() {
-        assert_eq!(
-            QueryResultMapper::map(Ok(QueryEvent::TableDropped)),
-            vec![Message::CommandComplete("DROP TABLE".to_owned())]
-        );
-    }
-
-    #[test]
-    fn insert_record() {
-        let records_number = 3;
-        assert_eq!(
-            QueryResultMapper::map(Ok(QueryEvent::RecordsInserted(records_number))),
-            vec![Message::CommandComplete(format!("INSERT 0 {}", records_number))]
-        )
-    }
-
-    #[test]
-    fn select_records() {
-        let projection = (
-            vec![
-                ("column_name_1".to_owned(), SqlType::SmallInt),
-                ("column_name_2".to_owned(), SqlType::SmallInt),
-            ],
-            vec![
-                vec!["1".to_owned(), "2".to_owned()],
-                vec!["3".to_owned(), "4".to_owned()],
-            ],
-        );
-        assert_eq!(
-            QueryResultMapper::map(Ok(QueryEvent::RecordsSelected(projection))),
-            vec![
-                Message::RowDescription(vec![
-                    ColumnMetadata::new("column_name_1".to_owned(), 21, 2),
-                    ColumnMetadata::new("column_name_2".to_owned(), 21, 2)
-                ]),
-                Message::DataRow(vec!["1".to_owned(), "2".to_owned()]),
-                Message::DataRow(vec!["3".to_owned(), "4".to_owned()]),
-                Message::CommandComplete("SELECT 2".to_owned())
-            ]
-        );
-    }
-
-    #[test]
-    fn update_records() {
-        let records_number = 3;
-        assert_eq!(
-            QueryResultMapper::map(Ok(QueryEvent::RecordsUpdated(records_number))),
-            vec![Message::CommandComplete(format!("UPDATE {}", records_number))]
-        );
-    }
-
-    #[test]
-    fn delete_records() {
-        let records_number = 3;
-        assert_eq!(
-            QueryResultMapper::map(Ok(QueryEvent::RecordsDeleted(records_number))),
-            vec![Message::CommandComplete(format!("DELETE {}", records_number))]
-        )
-    }
-
-    #[test]
-    fn schema_already_exists() {
-        let schema_name = "some_table_name".to_owned();
-        assert_eq!(
-            QueryResultMapper::map(Err(QueryError::schema_already_exists(schema_name.clone()))),
-            vec![Message::ErrorResponse(
-                Some("ERROR".to_owned()),
-                Some("42P06".to_owned()),
-                Some(format!("schema \"{}\" already exists", schema_name)),
-            )]
-        )
-    }
-
-    #[test]
-    fn schema_does_not_exists() {
-        let schema_name = "some_table_name".to_owned();
-        assert_eq!(
-            QueryResultMapper::map(Err(QueryError::schema_does_not_exist(schema_name.clone()))),
-            vec![Message::ErrorResponse(
-                Some("ERROR".to_owned()),
-                Some("3F000".to_owned()),
-                Some(format!("schema \"{}\" does not exist", schema_name)),
-            )]
-        )
-    }
-
-    #[test]
-    fn table_already_exists() {
-        let table_name = "some_table_name".to_owned();
-        assert_eq!(
-            QueryResultMapper::map(Err(QueryError::table_already_exists(table_name.clone()))),
-            vec![Message::ErrorResponse(
-                Some("ERROR".to_owned()),
-                Some("42P07".to_owned()),
-                Some(format!("table \"{}\" already exists", table_name)),
-            )]
-        )
-    }
-
-    #[test]
-    fn table_does_not_exists() {
-        let table_name = "some_table_name".to_owned();
-        assert_eq!(
-            QueryResultMapper::map(Err(QueryError::table_does_not_exist(table_name.clone()))),
-            vec![Message::ErrorResponse(
-                Some("ERROR".to_owned()),
-                Some("42P01".to_owned()),
-                Some(format!("table \"{}\" does not exist", table_name)),
-            )]
-        )
-    }
-
-    #[test]
-    fn one_column_does_not_exists() {
-        assert_eq!(
-            QueryResultMapper::map(Err(QueryError::column_does_not_exist(vec![
-                "column_not_in_table".to_owned()
-            ]))),
-            vec![Message::ErrorResponse(
-                Some("ERROR".to_owned()),
-                Some("42703".to_owned()),
-                Some("column column_not_in_table does not exist".to_owned()),
-            )]
-        )
-    }
-
-    #[test]
-    fn multiple_columns_does_not_exists() {
-        assert_eq!(
-            QueryResultMapper::map(Err(QueryError::column_does_not_exist(vec![
-                "column_not_in_table1".to_owned(),
-                "column_not_in_table2".to_owned()
-            ]))),
-            vec![Message::ErrorResponse(
-                Some("ERROR".to_owned()),
-                Some("42703".to_owned()),
-                Some("columns column_not_in_table1, column_not_in_table2 do not exist".to_owned()),
-            )]
-        )
-    }
-
-    #[test]
-    fn operation_is_not_supported() {
-        let raw_sql_query = "some SQL query".to_owned();
-        assert_eq!(
-            QueryResultMapper::map(Err(QueryError::not_supported_operation(raw_sql_query.clone()))),
-            vec![Message::ErrorResponse(
-                Some("ERROR".to_owned()),
-                Some("42601".to_owned()),
-                Some(format!("Currently, Query '{}' can't be executed", raw_sql_query)),
-            )]
-        )
     }
 }
