+++ src/sql_types/src/lib.rs
@@ -12,6 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use protocol::sql_types::PostgreSqlType;
 use serde::{Deserialize, Serialize};
 use std::convert::TryInto;
 
@@ -37,23 +38,43 @@ pub enum SqlType {
 impl SqlType {
     pub fn constraint(&self) -> Box<dyn Constraint> {
         match *self {
-            SqlType::Char(length) => Box::new(CharSqlTypeConstraint { length }),
-            SqlType::VarChar(length) => Box::new(VarCharSqlTypeConstraint { length }),
-            SqlType::SmallInt => Box::new(SmallIntTypeConstraint),
-            SqlType::Integer => Box::new(IntegerSqlTypeConstraint),
-            SqlType::BigInt => Box::new(BigIntTypeConstraint),
+            Self::Char(length) => Box::new(CharSqlTypeConstraint { length }),
+            Self::VarChar(length) => Box::new(VarCharSqlTypeConstraint { length }),
+            Self::SmallInt => Box::new(SmallIntTypeConstraint),
+            Self::Integer => Box::new(IntegerSqlTypeConstraint),
+            Self::BigInt => Box::new(BigIntTypeConstraint),
             sql_type => unimplemented!("Type constraint for {:?} is not currently implemented", sql_type),
         }
     }
 
     pub fn serializer(&self) -> Box<dyn Serializer> {
         match *self {
-            SqlType::Char(_length) => Box::new(CharSqlTypeSerializer),
-            SqlType::VarChar(_length) => Box::new(VarCharSqlTypeSerializer),
-            SqlType::SmallInt => Box::new(SmallIntTypeSerializer),
-            SqlType::Integer => Box::new(IntegerSqlTypeSerializer),
-            SqlType::BigInt => Box::new(BigIntTypeSerializer),
+            Self::Char(_length) => Box::new(CharSqlTypeSerializer),
+            Self::VarChar(_length) => Box::new(VarCharSqlTypeSerializer),
+            Self::SmallInt => Box::new(SmallIntTypeSerializer),
+            Self::Integer => Box::new(IntegerSqlTypeSerializer),
+            Self::BigInt => Box::new(BigIntTypeSerializer),
             sql_type => unimplemented!("Type Serializer for {:?} is not currently implemented", sql_type),
+        }
+    }
+
+    pub fn to_pg_types(&self) -> PostgreSqlType {
+        match *self {
+            Self::Bool => PostgreSqlType::Bool,
+            Self::Char(_) => PostgreSqlType::Char,
+            Self::VarChar(_) => PostgreSqlType::VarChar,
+            Self::Decimal => PostgreSqlType::Decimal,
+            Self::SmallInt => PostgreSqlType::SmallInt,
+            Self::Integer => PostgreSqlType::Integer,
+            Self::BigInt => PostgreSqlType::BigInt,
+            Self::Real => PostgreSqlType::Real,
+            Self::DoublePrecision => PostgreSqlType::DoublePrecision,
+            Self::Time => PostgreSqlType::Time,
+            Self::TimeWithTimeZone => PostgreSqlType::TimeWithTimeZone,
+            Self::Timestamp => PostgreSqlType::Timestamp,
+            Self::TimestampWithTimeZone => PostgreSqlType::TimestampWithTimeZone,
+            Self::Date => PostgreSqlType::Date,
+            Self::Interval => PostgreSqlType::Interval,
         }
     }
 }
@@ -216,6 +237,97 @@ impl Serializer for VarCharSqlTypeSerializer {
 #[cfg(test)]
 mod tests {
     use super::*;
+
+    #[cfg(test)]
+    mod to_postgresql_type_conversion {
+        use crate::SqlType;
+        use protocol::sql_types::PostgreSqlType;
+
+        #[test]
+        fn boolean() {
+            assert_eq!(SqlType::Bool.to_pg_types(), PostgreSqlType::Bool);
+        }
+
+        #[test]
+        fn small_int() {
+            assert_eq!(SqlType::SmallInt.to_pg_types(), PostgreSqlType::SmallInt);
+        }
+
+        #[test]
+        fn integer() {
+            assert_eq!(SqlType::Integer.to_pg_types(), PostgreSqlType::Integer);
+        }
+
+        #[test]
+        fn big_int() {
+            assert_eq!(SqlType::BigInt.to_pg_types(), PostgreSqlType::BigInt);
+        }
+
+        #[test]
+        fn char() {
+            assert_eq!(SqlType::Char(0).to_pg_types(), PostgreSqlType::Char);
+            assert_eq!(SqlType::Char(10).to_pg_types(), PostgreSqlType::Char);
+            assert_eq!(SqlType::Char(100).to_pg_types(), PostgreSqlType::Char);
+        }
+
+        #[test]
+        fn var_char() {
+            assert_eq!(SqlType::VarChar(0).to_pg_types(), PostgreSqlType::VarChar);
+            assert_eq!(SqlType::VarChar(10).to_pg_types(), PostgreSqlType::VarChar);
+            assert_eq!(SqlType::VarChar(100).to_pg_types(), PostgreSqlType::VarChar);
+        }
+
+        #[test]
+        fn decimal() {
+            assert_eq!(SqlType::Decimal.to_pg_types(), PostgreSqlType::Decimal);
+        }
+
+        #[test]
+        fn real() {
+            assert_eq!(SqlType::Real.to_pg_types(), PostgreSqlType::Real);
+        }
+
+        #[test]
+        fn double_precision() {
+            assert_eq!(SqlType::DoublePrecision.to_pg_types(), PostgreSqlType::DoublePrecision);
+        }
+
+        #[test]
+        fn time() {
+            assert_eq!(SqlType::Time.to_pg_types(), PostgreSqlType::Time);
+        }
+
+        #[test]
+        fn time_with_time_zone() {
+            assert_eq!(
+                SqlType::TimeWithTimeZone.to_pg_types(),
+                PostgreSqlType::TimeWithTimeZone
+            );
+        }
+
+        #[test]
+        fn timestamp() {
+            assert_eq!(SqlType::Timestamp.to_pg_types(), PostgreSqlType::Timestamp);
+        }
+
+        #[test]
+        fn timestamp_with_timezone() {
+            assert_eq!(
+                SqlType::TimestampWithTimeZone.to_pg_types(),
+                PostgreSqlType::TimestampWithTimeZone
+            );
+        }
+
+        #[test]
+        fn date() {
+            assert_eq!(SqlType::Date.to_pg_types(), PostgreSqlType::Date);
+        }
+
+        #[test]
+        fn interval() {
+            assert_eq!(SqlType::Interval.to_pg_types(), PostgreSqlType::Interval);
+        }
+    }
 
     #[cfg(test)]
     mod ints {
